+++ melting-java/src/melting/approximativeMethods/Ahsen01.java
@@ -19,7 +19,8 @@ public class Ahsen01 extends ApproximativeMode{
 	public ThermoResult calculateThermodynamics() {
 		double Tm = super.calculateThermodynamics().getTm();
 		double percentGC = this.environment.getSequences().calculatePercentGC();
-		Tm = 80.4 + 0.345 * percentGC + Math.log10(this.environment.getNa()) * (17.0 - 0.135 * percentGC) - 550.0 / this.environment.getSequences().getDuplexLength();
+		Tm = 80.4 + 0.345 * percentGC + Math.log10(this.environment.getNa()) * (17.0 - 0.135 * percentGC) - 550.0 / (double)this.environment.getSequences().getDuplexLength();
+
 		environment.setResult(Tm);
 		
 		OptionManagement.meltingLogger.log(Level.FINE, "from Ahsen et al. (2001)");
